requirements.txt
fastapi
uvicorn[standard]
python-dotenv
groq
python-docx
fpdf2
gunicorn